split: refactor filename suffix
